Move perf_timer into _timers.py
Remove _utils.py since there was only 1 thing left in it.
